+++ src/main/java/insane96mcp/iguanatweaksreborn/module/items/itemstats/ItemStats.java
@@ -175,6 +175,21 @@ public class ItemStats extends Feature {
 	}
 
 	@SubscribeEvent
+	public void onBlockRightClick(PlayerInteractEvent.RightClickBlock event) {
+		if (!this.isEnabled()
+				|| !unbreakableItems)
+			return;
+
+		ItemStack stack = event.getItemStack();
+		if (stack.getMaxDamage() == 0)
+			return;
+		if (isBroken(stack)) {
+			event.setCanceled(true);
+			event.getEntity().displayClientMessage(Component.translatable(BROKEN_ITEM_LANG), true);
+		}
+	}
+
+	@SubscribeEvent
 	public void onEntityInteract(PlayerInteractEvent.EntityInteract event) {
 		if (!this.isEnabled()
 				|| !unbreakableItems)